Move one more util function into it's rightful place
