+++ src/utils/convert.ts
@@ -403,3 +403,14 @@ export function calculateCropScales(value: number, isRaining: boolean) {
     cropRatio,
   };
 };
+
+
+export function convertDeltaDemandToPercentage(deltaDemand: number) {
+  if (deltaDemand === 0) return "0%";
+  if (deltaDemand === 1e18) return "100%";
+  if (deltaDemand === 1e36) return "∞%";
+
+  // Scale the value between 0-100%
+  const scaledValue = (deltaDemand / 1e18) * 100;
+  return `${TokenValue.fromHuman(scaledValue, 0).toHuman("short")}%`;
+}
